add PhpFunctionOverrides and rename HeaderFunctions to HeaderStack
