+++ composer.json
@@ -53,7 +53,7 @@
     },
     "autoload-dev": {
         "files": [
-            "tests/Assets/HeaderFunctions.php"
+            "tests/Assets/PhpFunctionOverrides.php"
         ]
     },
     "scripts": {
